requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-chess==1.999
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
stockfish==3.28.0
